Added part 2 and cleaned up unused methods

+++ advent_of_code_2021/day_16/utils.cpp
@@ -52,9 +52,9 @@ std::vector<bool> HexToBinary(std::string hexString)
     return v;
 }
 
-unsigned int BinaryToDecimal(const std::vector<bool> &binaryVector)
+unsigned long long BinaryToDecimal(const std::vector<bool> &binaryVector)
 {
-    unsigned int x{0};
+    unsigned long long x{0};
 
     size_t n{binaryVector.size()};
